# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start API",
        "description": "AgroBozor uchun backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.daladan.uz/api/v1/",
            "description": "Production server"
        },
        {
            "url": "http://daladan-api.loc/api/v1",
            "description": "Local dev server (OSPanel)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "php artisan serve"
        }
    ],
    "paths": {
...
        "/profile/ads/{ad}/promotions": {
            "post": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "E'longa promo buyurtmasi (pending)",
                "description": "Faqat o'z e'loniga. Keyin to'lov + admin PATCH /admin/ad-promotions/{promotion}/confirm.",
                "operationId": "85256a9961dded30cd312dc050633cc0",
                "parameters": [
                    {
                        "name": "ad",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "promotion_plan_id"
                                ],
                                "properties": {
                                    "promotion_plan_id": {
                                        "description": "GET /resources/promotion-plans",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Buyurtma yaratildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdPromotionStoreSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "E'lon topilmadi"
                    },
                    "422": {
                        "description": "Validatsiya yoki biznes qoidasi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/public/ads": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Barcha faol e'lonlar (ixtiyoriy filter)",
                "description": "category_id/subcategory_id yuborilmasa — barcha faol e'lonlar; yuborilsa — shu bo'yicha filter.",
                "operationId": "728b62ca4aa4bee130a611e83c0407cd",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "1–50, default 15",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Berilmasa barcha kategoriyalar",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subcategory_id",
                        "in": "query",
                        "description": "Berilmasa filter yo'q",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginatsiyali e'lonlar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicAdListSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "category_id/subcategory_id noto'g'ri"
                    }
                }
            }
        },
        "/public/ads/{id}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Bitta faol e'lon",
                "description": "ad() — GET /public/ads/{id}",
                "operationId": "3dedff3ec13a5e42171cd56f935d0232",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "E'lon obyekti",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicAdSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Topilmadi yoki faol emas"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdPromotionPlanSummary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Top sotuv — 7 kun"
                    },
                    "slug": {
                        "type": "string",
                        "example": "top-sale-7d"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "top_sale",
                            "boost"
                        ]
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 9000
                    },
                    "currency": {
                        "type": "string",
                        "example": "UZS"
                    },
                    "duration_days": {
                        "type": "integer",
                        "example": 7
                    }
                },
                "type": "object"
            },
            "AdPromotionOrderItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ad_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "promotion_plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 9000
                    },
                    "currency": {
                        "type": "string",
                        "example": "UZS"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "paid",
                            "active",
                            "expired",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "payment_transaction_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "promotion_plan": {
                        "$ref": "#/components/schemas/AdPromotionPlanSummary"
                    }
                },
                "type": "object"
            },
            "AdPromotionStoreSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AdPromotionOrderItem"
                    }
                },
                "type": "object"
            },
...
            "PublicAd": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 201
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "subcategory_id": {
                        "type": "integer",
                        "example": 11
                    },
                    "seller_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Naslli echkilar"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "district": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "integer",
                        "example": 2200000,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "number",
                        "format": "float",
                        "example": 12,
                        "nullable": true
                    },
                    "unit": {
                        "type": "string",
                        "example": "piece",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "is_top_sale": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_boosted": {
                        "type": "boolean",
                        "example": false
                    },
                    "boost_starts_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "boost_expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "highlight_active": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PublicAdSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PublicAd"
                    }
                },
                "type": "object"
            },
            "PublicAdListSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 300
                            },
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PublicAd"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}